+++ api/api_router.js
@@ -3,8 +3,8 @@ const restricted = require('../middleware/restricted_middleware');
 const authRouter = require('../auth/auth_router');
 const userRouter = require('../users/users_router');
 const video_update = require('../videos/video_update');
-const videos = require('../videos/videos');
-const songs = require('../songs/songs');
+const videos = require('../videos/videos_router');
+const songs = require('../songs/songs_router');
 
 router.use('/auth', authRouter);
 router.use('/users', restricted, userRouter);
